tests/wpt/web-platform-tests/svg/painting/parsing/marker-end-valid.svg

<?xml version="1.0" encoding="UTF-8"?>
<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml"
     width="800px" height="800px">
  <title>SVG Painting: parsing marker-end with valid values</title>
  <metadata>
    <h:link rel="help" href="https://svgwg.org/svg2-draft/painting.html#MarkerEndProperty"/>
    <h:meta name="assert" content="marker-end supports the full paint grammar 'none | marker-ref'."/>
  </metadata>
  <g id="target"></g>
  <h:script src="/resources/testharness.js"/>
  <h:script src="/resources/testharnessreport.js"/>
  <h:script src="/css/support/parsing-testcommon.js"/>
  <script><![CDATA[

test_valid_value("marker-end", "none");

test_valid_value("marker-end", 'url("https://example.com/")');

  ]]></script>
</svg>
